+++ MarkdownPreviewUITests/MarkdownPreviewUITestsLaunchTests.swift
@@ -13,10 +13,6 @@ import XCTest
 
 final class MarkdownPreviewUITestsLaunchTests: XCTestCase {
 
-    override class var runsForEachTargetApplicationUIConfiguration: Bool {
-        true
-    }
-
     override func setUpWithError() throws {
         continueAfterFailure = false
     }